Started to add astGetRegionBounds.
This allows the rest of the STCS test from Pyast to be
performed, as well as allowing the bounds to be set
for plotting of regions.

At the moment it just pushes the return values into the
arrays specified, and should be updated to clear them
first.  (Or to create arrays and return them.)