Fix "attempt to free unreferenced scalar" warnings
These occured when the module croaked with an AST error. There was
a problem with the mortalisation of the elements of the error
stack where I was making elements of the AV mortal instead of just
the AV itself. Now we just store the SV*s from the global stack
directly in the local AV* and just increment the reference counter
whilst also clearin the global stack.

This patch includes a test and a cleanup of the exception triggering
code so that we no longer call back into perl simply to create
a string and croak.