Include AST source code as a submodule
This includes the AST source code as "extsrc".  A new build action,
"prepare" was added as a pre-requisite of the "code" and "dist"
actions.  This bootstraps and configures the "extsrc" directory
and makes some headers.  Files which were previously included
are then copied to the "ast" directory so as to re-generate the
original directory structure.

This means that a Starlink environment is needed initially to make
a package for distribution.  However the contents of that package
should be the same as before.  (And the "prepare" action does nothing
when the "ast" directory already exists.)