Add support for new MOC and related functions
The "ast_tester/testmoc.f" test program was converted to Perl to
create a new test script "t/moc.t".  The functions required to
run this test script where then added to the wrapper.

Some functions which use 2D arrays are currently exchanging these
with Perl as 1D arrays since that is what is supported by the
"arrays.c" library.