Remove myAstRate wrapper function
Since we no longer have version-specific functionality, the
separate myAstRate wrapper function (implementing astRate) is not
required for the purpose of allowing preprocessor directives.
With this wrapper function in place the output value seemed
never to make it back to Perl routines calling astRate.
Add a test script with conversions to Perl of a couple of AST's
tests for astRate in order to ensure the function is returning
values properly.